requirements.txt
boto3==1.35.3
flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1